requirements.txt
annotated-types==0.6.0
appnope==0.1.4
asttokens==2.4.1
beautifulsoup4==4.12.3
blinker==1.8.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
debugpy==1.8.1
decorator==5.1.1
exceptiongroup==1.2.1
executing==2.0.1
Flask==3.0.3
google-ai-generativelanguage==0.6.2
google-api-core==2.18.0
google-api-python-client==2.127.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.2
googleapis-common-protos==1.63.0
grpcio==1.62.2
grpcio-status==1.62.2
httplib2==0.22.0
idna==3.7
importlib_metadata==7.1.0
iniconfig==2.0.0
ipykernel==6.29.4
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.1
jupyter_core==5.7.2
lxml==5.2.1
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
ordereddict==1.1
packaging==24.0
parso==0.8.4
pep8==1.7.1
pexpect==4.9.0
platformdirs==4.2.1
pluggy==1.5.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.17.2
pyparsing==3.1.2
pyrate-limiter==3.6.0
pytest==8.1.2
python-dateutil==2.9.0.post0
python-xbrl==1.1.1
pyzmq==26.0.2
requests==2.31.0
rsa==4.9
sec-edgar-downloader==5.0.2
six==1.16.0
soupsieve==2.5
stack-data==0.6.3
tomli==2.0.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.3
zipp==3.18.1
